backend/requirements.txt
# web API
fastapi>=0.115,<0.116
uvicorn[standard]>=0.30,<0.31

# models & validation
pydantic>=2.7,<3

# config, http
python-dotenv>=1.0,<2
httpx>=0.27,<0.28
tenacity>=8.2,<9

# AI services
openai>=1.44,<2

# firestore (server-side)
google-cloud-firestore>=2.16,<3
google-auth>=2.32,<3
google-auth-oauthlib>=1.2,<2
